### 2022-10-28 #4888 @editorialbot
[PRE REVIEW]: Lightshow: a Python package for generating computational x-ray absorption spectroscopy input files
**Submitting author:** @x94carbone (<a href="http://orcid.org/0000-0002-5181-9513">Matthew Carbone</a>)
**Repository:** https://github.com/AI-multimodal/Lightshow
**Branch with paper.md** (empty if default branch):
**Version:** v0.0.14
