### 2022-10-28 18ce049
Update README.rst
Add JOSS badge
