+++ pyproject.toml
@@ -40,7 +40,7 @@ dependencies = [
     "numpy",
     "tqdm",
     "monty",
-    "pymatgen",
+    "pymatgen<=2022.7.8",
     "ase"
 ]
 
