### 2022-10-30 #4888 @x94carbone
Editors, I hot fixed the install to version-lock the Pymatgen dependency, since somehow a more recent version of Pymatgen breaks only some of the materials in our CI pipeline. New version is `v0.0.15`.
