### 2022-11-01 #4888 @Kevin-Mattheus-Moerman
@x94carbone thanks for this submission. I have just added the query scope label here. This means our editorial board will review if this work is in scope for JOSS. Although extensive enough in terms of lines of code (we set a rough bar at around 1000 lines of code) this submission is on the small side. Furthermore, as can be seen in our [submission requirements](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort), "single functionality" packages are usually not in scope. From studying your repository and paper it looks to me like this work is a "generator of input files", as such this software perhaps classifies as a "single functionality" package. It does look like it is quite a technical challenge that is being solved, as generating these files sounds non-trivial. Hence I need the help of the board to review this in a bit more detail. Feel free to add some comments here too if you'd like to further clarify how this software goes beyond offering a "single functionality".
The scope review should take about 1 week to complete and we will get back to you shortly here.
