### 2022-11-01 #4888 @x94carbone
@Kevin-Mattheus-Moerman sure no problem. Yep I read the submission guidelines before submitting and I noticed that we are slightly below the 1000 line standard requirement (CodeCov at least says 912) and thought we should take our chances since we're well above the 300 line desk-reject limit. I am happy to provide additional justification of why this belongs in JOSS.

> From studying your repository and paper it looks to me like this work is a "generator of input files", as such this software perhaps classifies as a "single functionality" package.

While it is true that Lightshow is designed for writing computational spectroscopy input files, I want to highlight a few features that put this outside of the "single functionality" box, at least in my opinion.

1.  If we only wrote input files for one type of code and did not offer a uniform abstraction for easy extendability, then I would agree it's probably too simple. However, Lightshow currently can write input files for five different types of x-ray spectroscopy softwares (as noted in the manuscript, these are FEFF, VASP, EXCITING, XSPECTRA and OCEAN). It can (and does) also do other things, which I detail later.
2. We also handle an annoying "atom alignment" problem for those users who wish to do e.g. benchmarking. In other words, despite the fact that some codes are "cluster-based" and some are "supercell-based" (and different codes order atoms differently in their input files) every input file for material `X` and atom `N` for FEFF calculations will be the "same atom" in every other type of calculation. This is extensively tested and I am not aware of another software that unifies this process, which is key for generating data for benchmarking or data-driven purposes when multiple "levels of theory" are required.

So I suppose that our code offers at least two key functionalities. Depending on how you define functionalities, I think I could also argue (as I will later down below) that we have at least five key functionalities given Lightshow has the ability to write input files for five different simulation codes, and each comes with its own unique challenges.

You are correct that the current scope of this code is to write computational spectroscopy input files, but it can do this in many ways depending on what users want to do. For example, key functionality is exposed for every one of the types of codes we write input files for that allow users to tune code-specific (VASP potential files) and code-agnostic properties (the number of bands in a supercell calculation). Users can decide to pull materials data from the Materials Project, or to load data manually from their own hard drive. Etc.

Additionally, the code does (in my opinion, clearly) meet all of the other requirements, including feature completeness, packaged according to community, etc. Importantly, we are also designed to be extendable. Lightshow does not just need to write computational spectroscopy input files. In fact, for users' convenience in calculating the appropriate energy shifts between a neutral and core-hole calculation, we make it easy to run ground state convergence calculations (non-spectroscopy) in VASP, and some of the other codes. If this is too domain-specific an explanation, think "we also do other non-spectroscopy calculations that users might need in their analysis of the spectroscopy results".

Finally, I would ask that the editors consider that significant effort went into "golfing the code" down from bloated to what it is now. The implementation of a uniform abstraction for writing the input files saved us considerable space. So while the code is ~900 lines, it is a dense 900 lines that started from many more. We could've left it as-is and it probably still would've worked fine, but the team decided it would be worth the effort to consolidate since that would make the software much more maintainable.

> It does look like it is quite a technical challenge that is being solved, as generating these files sounds non-trivial.

Thank you, indeed it definitely was.

Point 1 above for especially VASP, EXCITING, XSPECTRA and OCEAN was a huge challenge to get right. We actually have core developers from EXCITING and OCEAN on this team (Christian, Claudia and John), and experts on the other three codes, to ensure that we did.

Point 2 ("atom alignment") was another very annoying problem we needed to solve robustly in order to benchmark these codes against each other and generate machine learning databases for ongoing work. Due to the nuances of each individual code this also proved to be very challenging. It's also a pretty easy pitfall to get stuck in (speaking from experience here), which we abstract away for the user.

> The scope review should take about 1 week to complete and we will get back to you shortly here.

Thank you for your time. I would only ask on behalf of our team that the editors consider the above. I do believe that Lightshow is multifunctional, and it was designed for extendability so that more functionalities can be easily added by the community. Please do feel free to ask any questions should you have them.
