### 2022-11-02 #4888 @Kevin-Mattheus-Moerman
@x94carbone we have reviewed your work and your comments and feel it may be in scope for JOSS. I have just removed the query-scope label. However I may need to add the `waitlisted` label now since the editors in this domain are currently handling many other submissions. We will assign an editor once one becomes available.
