### 2022-11-07 #4888 @ppxasjsm
@x94carbone, I'll take this paper on as an editor. I do have some other work on my plate right now, but I'll do my best to process this speedily!

If you have any suggestions for potential reviewers then please mention them here in this thread (without tagging them with an @). In addition, [this list of people](https://bit.ly/joss-reviewers) have already agreed to review for JOSS and may be suitable for this submission (please start at the bottom of the list). This will speed up the process for me to identify suitable reviewers.
