### 2022-11-08 #4888 @x94carbone
@ppxasjsm ok some possible reviewers:

- maurov lists x-ray spectroscopy as a specialty along with Python, which is basically a perfect fit for our work
- marwanalam2 actually lists VASP as one of their programming languages (not sure I'd agree it's a language but hey why not?). VASP is one of the codes that we hook into in Lightshow
- domna, larsenkg, enricgrau and tillsten are all pretty far down the list and list spectroscopy as a preferred area (and seemingly the types of spectroscopy that could be relevant to this work)

I believe the first two would be the best fits, probably could start there!
