### 2022-11-24 #4888 @maurov
> @maurov, I was wondering if you'd be willing to review this paper for me?

@ppxasjsm yes, this paper fits my domain of expertise and I am happy to review it. Please, let me know the steps to follow, as this is my first review for JOSS.
