### 2022-11-28 #4888 @ppxasjsm
Brilliant, thank you @maurov. For now, I'll add you as a reviewer. I am still looking for a second reviewer, so it will take a little moment before more reviewing instructions will follow. In the meantime you can have a look [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html).
