### 2022-12-08 c060da0
Fixed code formatting errors to pass Flake8 and Black tests
