### 2022-12-09 9cc7e06
Delete CITATION.cff
Remove the CITATION file now that we have an arXiv preprint
