### 2023-02-21 #5182 @editorialbot
[REVIEW]: Lightshow: a Python package for generating computational x-ray absorption spectroscopy input files
**Submitting author:** @matthewcarbone (<a href="http://orcid.org/0000-0002-5181-9513">Matthew Carbone</a>)
**Repository:** https://github.com/AI-multimodal/Lightshow
**Branch with paper.md** (empty if default branch): joss-manuscript
**Version:** v0.1.3
