+++ paper.bib
@@ -16,7 +16,8 @@
   number={21},
   pages={5503--5513},
   year={2010},
-  publisher={Royal Society of Chemistry}
+  publisher={Royal Society of Chemistry},
+  doi={https://doi.org/10.1039/B926434E}
 }
 
 @article{taillefumier2002x,
@@ -60,7 +61,8 @@
   number={11},
   pages={115106},
   year={2011},
-  publisher={APS}
+  publisher={APS},
+  doi={https://doi.org/10.1103/PhysRevB.83.115106}
 }
 
 @article{ocean-3,
@@ -90,7 +92,8 @@
   journal={Electronic Structure},
   volume={1}, 
   pages={037001},
-  year={2019}
+  year={2019},
+  doi={https://doi.org/10.1088/2516-1075/ab3123}
 }
 
 @article{xanesbench,
