+++ paper.bib
@@ -17,7 +17,7 @@
   pages={5503--5513},
   year={2010},
   publisher={Royal Society of Chemistry},
-  doi={doi.org/10.1039/B926434E}
+  doi={10.1039/B926434E}
 }
 
 @article{taillefumier2002x,
@@ -62,7 +62,7 @@
   pages={115106},
   year={2011},
   publisher={APS},
-  doi={doi.org/10.1103/PhysRevB.83.115106}
+  doi={10.1103/PhysRevB.83.115106}
 }
 
 @article{ocean-3,
@@ -93,7 +93,7 @@
   volume={1}, 
   pages={037001},
   year={2019},
-  doi={doi.org/10.1088/2516-1075/ab3123}
+  doi={10.1088/2516-1075/ab3123}
 }
 
 @article{xanesbench,
