### 2023-02-21 #5182 @matthewcarbone
@ppxasjsm I've added the DOIs explicitly for those manuscripts. Should I add every DOI? Seems like it's imbalanced now. Certain references have DOIs and some don't!
