### 2023-02-23 #5182 @maurov
@matthewcarbone I have started the review and my *main concern* is that I do not understand why you have decided to create a new software from scratch instead contributing to existing well established open-source Python projects in the field, namely, [Pymatgen](https://pymatgen.org/) and [ASE](https://wiki.fysik.dtu.dk/ase/)?

Since you are closer to Pymatgen, why not simply extending  [pymatgen.io](https://pymatgen.org/pymatgen.io.html) with OCEAN, EXCITING and Xspectra?

Furthermore, in the statement of need is written that Lightshow is a tool for XAS simulation and **analysis**. In the current status, I see only a (possibly redundant) tool for generating XAS input files, but I do not see any tool helpful for the analysis of the XAS simulations. As part of the XAS community myself, I would expect much more on the analysis side. For example: a tool for extracting local coordination environment from the input structures or modify the local environment (e.g. angles, coordination, distances); perform convergence tests versus a given structural parameter; handle the generated spectra (e.g. plots and validation/comparison with experimental data).

I think that all this should be clearly stated out in the software paper and in the documentation.

The current documentation is minimalist and does not show a single case of the full workflow: from the input structure down to a XAS spectrum. I think this should go into the quickstart or refer to a more complete notebook (just generating the input file is not enough). In my opinion, the first/minimal thing a generic user would like to do with Lightshow is: 1) load a CIF file, 2) generate an input file for a given calculator, 3) run the calculator, 4) plot the XAS spectrum. Points 3 and 4 are very important in the workflow.

I do think there is a need in the scientific community for such a tool and your research group has the expertise for it, but with the current contribution you are sharing only the tip of the iceberg. So, please, add more detailed notebooks with real cases examples.

I would also describe in the documentation how to extend Lightshow with other calculators or add at least input file generation for [FDMNES](https://fdmnes.neel.cnrs.fr/), which is well known in the XAS community.

These are my general comments. I will add more technical aspects later on or directly link to Github issues.
