### 2023-02-23 #5182 @matthewcarbone
@maurov thanks for the initial feedback. I can try to address some of your concerns here, and I'll get in touch with the team to handle the more technical comments regarding e.g. notebooks.

> I do not understand why you have decided to create a new software from scratch instead contributing to existing well established open-source Python projects in the field, namely, [Pymatgen](https://pymatgen.org/) and [ASE](https://wiki.fysik.dtu.dk/ase/)?
> Since you are closer to Pymatgen, why not simply extending [pymatgen.io](https://pymatgen.org/pymatgen.io.html) with OCEAN, EXCITING and Xspectra?

In a perfect world, this is the correct open-source solution, but unfortunately there are fundamental limitations we are bound by at my institution. For one, this work was funded by members of BNL and therefore we are tied down by certain requirements. For example, a full "project," if you will, is required to be licensed in a certain way. Making these contributions upstream to `ase` or `pymatgen` was unfortunately just not an option for us. That said, now that our code is published, it is certainly permissible for the maintainers of these packages (or honestly anyone, including us) to copy/paste Lightshow code upstream as long as long as they copy/paste the license with it!

> Furthermore, in the statement of need is written that Lightshow is a tool for XAS simulation and analysis. In the current status, I see only a (possibly redundant) tool for generating XAS input files, but I do not see any tool helpful for the analysis of the XAS simulations.

I don't think it's redundant. It is _slightly_ redundant for FEFF but we've not found a tool that does what we want for the other codes (Pymatgen can handle VASP to some extent but it's very confusing, at least to me). For benchmarking specifically we offer some functionality that is uniquely useful (matching unique sites between different codes). That said, analysis tools are _absolutely_ the next step and are definitely going to be a very useful addition for us and others.

> As part of the XAS community myself, I would expect much more on the analysis side. For example: a tool for extracting local coordination environment from the input structures or modify the local environment (e.g. angles, coordination, distances); perform convergence tests versus a given structural parameter; handle the generated spectra (e.g. plots and validation/comparison with experimental data).

All great ideas that I agree with. We just felt it was out-of-scope for the "`v1`" release.

> In my opinion, the first/minimal thing a generic user would like to do with Lightshow is: 1) load a CIF file, 2) generate an input file for a given calculator, 3) run the calculator, 4) plot the XAS spectrum. Points 3 and 4 are very important in the workflow.

At its current stage, Lightshow is not a workflow code. The subtleties in developing workflows for e.g. different computer architectures was just not something we wanted to address right now. Even `ase` often just requires an executable in the PATH, and then just calls that executable from a lightweight Python function. We figured at this stage that part was not really required for what we want to do.

Lightshow's intention at this stage is to write the input files for the user. Running the code is a whole different can of worms that we agreed was not in scope for the first release. That said, I think one thing we can do at least is to make a tutorial of how you'd run the codes using a simple `bash` script on some HPC, or something like this.

> I do think there is a need in the scientific community for such a tool and your research group has the expertise for it, but with the current contribution you are sharing only the tip of the iceberg. So, please, add more detailed notebooks with real cases examples.

Will do!

> I would also describe in the documentation how to extend Lightshow with other calculators or add at least input file generation for [FDMNES](https://fdmnes.neel.cnrs.fr/), which is well known in the XAS community.

FDMNES is actually at the top of our list for the next kind of code to add. We definitely are seeing it all over the place and agree it makes sense to add to Lightshow.
