+++ lightshow/_tests/test_database.py
@@ -17,7 +17,7 @@ from lightshow.defaults import VASP_INCAR_DEFAULT_COREHOLE_POTENTIAL
 
 # Helper testing files
 sys.path.append(str(Path(__file__).parent.resolve() / Path("helpers")))
-from geometry import consistency_check  # noqa
+from geometry import consistency_check  # noqa  # type: ignore
 
 
 def test_database_from_disk(database_from_file, test_structure_names):
@@ -65,6 +65,7 @@ def test_write(
     dat = database_from_file
     dat.initialize_supercells(9.0)
     dat.initialize_inequivalent_sites()
+    print(dat.structures.keys())
     dat._supercells = {mpid: dat.supercells[mpid]}
     dat._structures = {mpid: dat.structures[mpid]}
     dat._metadata = {mpid: dat.metadata[mpid]}
