### 2023-03-06 3a34eb9
Add python 3.10 to the list of versions
