### 2023-03-06 9c4cf07
Fix fragile tests path resolution issue
