### 2023-04-06 #5182 @maurov
@matthewcarbone regarding the **software paper**, I think overall reads well. I recommend adding a couple of sentences about the fact that this code is part of a (much) larger workflow consisting in running the ab-initio codes, generating the simulated spectra and comparing/benchmarking with experimental spectra.

Please, could you consider the following points ?

- correct citation syntax for [Meng et al.](https://doi.org/10.48550/arXiv.2303.17089)
- add citation for [Pymatgen](https://doi.org/10.1016/j.commatsci.2012.10.028)
- personally, I would add a citation for [FEFF10/Corvus](https://journals.iucr.org/s/issues/2021/06/00/hf5422/)
