### 2023-04-08 a115ddf
remove nkpt
pull request for ocean jupyternote book only
