+++ lightshow/parameters/ocean.py
@@ -313,7 +313,6 @@ class OCEANParameters(MSONable, _BaseParameters):
             # Standardized methods for getting the kpoints
             kmesh = self._kpoints(structure)
             cards["ngkpt"] = f"{kmesh[0]} {kmesh[1]} {kmesh[2]}"
-            cards["nkpt"] = f"{kmesh[0]} {kmesh[1]} {kmesh[2]}"
 
             # Determine the diemac
             if self._bandgap is not None or self._diel is not None:
