### 2023-04-12 #5182 @matthewcarbone
@maurov thanks for the additional feedback! We'll update the manuscript with your suggestions. In addition, the tutorial notebook is now much more comprehensive. [Here](https://github.com/AI-multimodal/Lightshow/tree/6cd9bcec4ea8ba775ae706f11648dcc8f0fa3bda) is the link to the commit, but it's currently on `master`.
