### 2023-04-12 AI-multimodal/Lightshow#169 @matthewcarbone
Address documentation concerns
Minor changes requested during JOSS review. See [here](https://github.com/openjournals/joss-reviews/issues/5182#issuecomment-1505563873).
