### 2023-04-13 b40aa03
Update pyproject.toml
Remove codecov from dependencies
