+++ lightshow/_tests/sample_files/README.rst
@@ -1,0 +1,1 @@
+.. sample broadened ti k-edge anatase spectra using different method as reference
