+++ notebooks/spectra_files/anatase_exp.txt
@@ -1,0 +1,112 @@
+4965.268571428572    0.04018369690011481    
+4965.611428571428    0.05281285878300804    
+4965.954285714286    0.06199770378874857    
+4966.125714285714    0.08036739380022961    
+4966.434285714286    0.09414466130884042    
+4966.88    0.11021814006888633    
+4967.702857142857    0.11940298507462686    
+4968.114285714286    0.10677382319173363    
+4968.765714285714    0.08955223880597014    
+4969.314285714286    0.09184845005740529    
+4969.657142857143    0.10562571756601608    
+4970.205714285714    0.11251435132032146    
+4970.48    0.11251435132032146    
+4971.062857142857    0.1044776119402985    
+4971.542857142857    0.0975889781859931    
+4971.92    0.09184845005740529    
+4972.194285714286    0.09184845005740529    
+4972.5371428571425    0.09184845005740529    
+4973.051428571429    0.11021814006888633    
+4973.291428571429    0.12169919632606199    
+4973.668571428571    0.1377726750861079    
+4973.908571428571    0.1584385763490241    
+4974.114285714286    0.17451205510907003    
+4974.354285714286    0.19288174512055106    
+4974.5257142857145    0.2123995407577497    
+4974.731428571428    0.22732491389207807    
+4974.937142857143    0.24339839265212398    
+4975.177142857143    0.2663605051664753    
+4975.348571428572    0.2835820895522388    
+4975.451428571429    0.30998851894374285    
+4975.76    0.3329506314580941    
+4975.862857142857    0.35706084959816303    
+4976.137142857143    0.3857634902411021    
+4976.274285714286    0.40642939150401836    
+4976.5142857142855    0.4225028702640643    
+4977.097142857143    0.4305396096440872    
+4977.645714285714    0.4420206659012629    
+4978.125714285714    0.4661308840413318    
+4978.674285714285    0.49598163030998854    
+4979.017142857143    0.5235361653272101    
+4979.325714285715    0.5464982778415615    
+4979.4628571428575    0.5671641791044776    
+4979.702857142857    0.5878300803673938    
+4979.908571428571    0.6073478760045924    
+4980.182857142857    0.6280137772675086    
+4980.457142857143    0.6486796785304247    
+4980.56    0.6659012629161882    
+4980.765714285714    0.6888633754305395    
+4980.902857142857    0.711825487944891    
+4981.108571428572    0.7324913892078071    
+4981.142857142857    0.7543053960964409    
+4981.28    0.7715269804822042    
+4981.417142857143    0.7956371986222732    
+4981.52    0.8151549942594719    
+4981.657142857143    0.8335246842709528    
+4981.794285714286    0.8473019517795637    
+4981.931428571429    0.8679678530424799    
+4982.171428571429    0.8886337543053962    
+4982.411428571429    0.9058553386911595    
+4982.617142857142    0.9207807118254879    
+4983.028571428571    0.9288174512055108    
+4983.5085714285715    0.9288174512055108    
+4983.851428571428    0.912743972445465    
+4984.228571428572    0.9012629161882892    
+4984.5371428571425    0.9012629161882892    
+4985.017142857143    0.9024110218140069    
+4985.36    0.9138920780711826    
+4985.668571428571    0.9276693455797934    
+4986.114285714286    0.9460390355912743    
+4986.4914285714285    0.9598163030998852    
+4986.834285714286    0.9793340987370838    
+4987.108571428572    0.9919632606199771    
+4987.691428571428    1.0    
+4988.548571428571    0.9954075774971296    
+4989.165714285714    0.9839265212399539    
+4989.337142857143    0.9632606199770378    
[…]
