+++ lightshow/postprocess/broaden.py
@@ -15,23 +15,23 @@ def gauss_broaden(x, xin, yin, sigma, shift=0.0):
 
     Parameters
     ----------
-    x: np.ndarray
+    x : numpy.ndarray
         The output energy grid.
-    xin: np.ndarray
+    xin : numpy.ndarray
         The input energy grid.
-    yin: np.ndarray
-        The input spectrum
-    sigma: float
-        Gaussian broadening sigma
-    shift: float
-        The energy shift
+    yin : numpy.ndarray
+        The input spectrum.
+    sigma : float
+        Gaussian broadening sigma.
+    shift : float
+        The energy shift.
 
     Returns
     -------
-    np.ndarray
+    numpy.ndarray
         Broadened spectrum
-
     """
+
     x1, x2 = np.meshgrid(x, xin + shift)
 
     return (
@@ -44,23 +44,23 @@ def lorentz_broaden(x, xin, yin, gamma, shift=0.0):
 
     Parameters
     ----------
-    x: np.ndarray
+    x : numpy.ndarray
         The output energy grid.
-    xin: np.ndarray
+    xin : numpy.ndarray
         The input energy grid.
-    yin: np.ndarray
-        The input spectrum
-    gamma: float
-        Lorentzian broadening gamma
-    shift: float
-        The energy shift
+    yin : numpy.ndarray
+        The input spectrum.
+    gamma : float
+        Lorentzian broadening gamma.
+    shift : float
+        The energy shift.
 
     Returns
     -------
-    np.ndarray
-        Broadened spectrum
-
+    numpy.ndarray
+        Broadened spectrum.
     """
+
     x1, x2 = np.meshgrid(x, xin + shift)
 
     return (
@@ -73,25 +73,25 @@ def voigt_broaden(x, xin, yin, sigma, gamma, shift=0.0):
 
     Parameters
     ----------
-    x: np.ndarray
+    x : numpy.ndarray
         The output energy grid.
-    xin: np.ndarray
+    xin : numpy.ndarray
         The input energy grid.
-    yin: np.ndarray
-        The input spectrum
-    sigma: float
-        Voigt broadening sigma
-    gamma: float
-        Voigt broadening gamma
-    shift: float
-        The energy shift
+    yin : numpy.ndarray
+        The input spectrum.
+    sigma : float
+        Voigt broadening sigma.
+    gamma : float
+        Voigt broadening gamma.
+    shift : float
+        The energy shift.
 
     Returns
     -------
-    np.ndarray
-        Broadened spectrum
-
+    numpy.ndarray
+        Broadened spectrum.
     """
+
     x1, x2 = np.meshgrid(x, xin + shift)
 
     return (
@@ -102,22 +102,19 @@ def voigt_broaden(x, xin, yin, sigma, gamma, shift=0.0):
 
 
 class SiteSpectum:
-    """
-    The class that store information about an XAS spectrum at each site,
-    including spectrum itself and the fermi energy, site weight.
+    """This class stores information about an XAS spectrum at each site,
+    including the spectrum itself, the Fermi energy and site weight.
 
     Parameters
     ----------
-    xin: np.ndarray
+    xin : numpy.ndarray
         The input energy grid.
-    yin: np.ndarray
-        The input spectrum
-
-    fermi: float
-        The fermi energy of the spectrum
-
-    weight: int
-        The weight of the site spectrum
+    yin : numpy.ndarray
+        The input spectrum.
+    fermi : float
+        The fermi energy of the spectrum.
+    weight : int
+        The weight of the site spectrum.
     """
 
     def __init__(
@@ -146,9 +143,8 @@ class SiteSpectum:
 
 
 class Broaden:
-    """
-    The class that store broadening parameters and provide functions to optimize
[…]
