### 2023-04-18 4d59d3c
crystal size uses Bohr as unit
