+++ lightshow/postprocess/broaden.py
@@ -1,30 +1,34 @@
 """XAS broadening using manual or empirically optimized parameters"""
 
-
-from scipy.stats import pearsonr as ps
+from monty.json import MSONable
+import numpy as np
+from scipy.stats import pearsonr
 from scipy import optimize
+from scipy.interpolate import InterpolatedUnivariateSpline
 from scipy.stats import norm
 from scipy.stats import cauchy
 from scipy.special import voigt_profile as voigt
 
-import numpy as np
 
+def gauss_broaden(xout, xin, yin, sigma):
+    """Applies Gaussian broadening to the input data, with standard deviation
+    sigma.
 
-def gauss_broaden(x, xin, yin, sigma, shift=0.0):
-    """Gaussian broadening function
+    .. math::
+
+        G(x; \\sigma) = \\frac{1}{\\sqrt{2 \\pi \\sigma^2}}
+        e^{-x^2 / 2\\sigma^2}
 
     Parameters
     ----------
-    x : numpy.ndarray
-        The output energy grid.
+    xout : numpy.ndarray
+        The output x-axis grid.
     xin : numpy.ndarray
-        The input energy grid.
+        The input x-axis grid.
     yin : numpy.ndarray
-        The input spectrum.
+        The input data for f(x) to broaden. Generally a spectrum.
     sigma : float
-        Gaussian broadening sigma.
-    shift : float
-        The energy shift.
+        The standard deviation of the Gaussian function used for broadening.
 
     Returns
     -------
@@ -32,14 +36,12 @@ def gauss_broaden(x, xin, yin, sigma, shift=0.0):
         Broadened spectrum
     """
 
-    x1, x2 = np.meshgrid(x, xin + shift)
-
-    return (
-        np.dot(norm.pdf(x1, x2, sigma).T, yin) / len(xin) * (xin[-1] - xin[0])
-    )
+    x1, x2 = np.meshgrid(xout, xin)
+    dx = xin[-1] - xin[0]
+    return np.dot(norm.pdf(x1, x2, sigma).T, yin) / len(xin) * dx
 
 
-def lorentz_broaden(x, xin, yin, gamma, shift=0.0):
+def lorentz_broaden(x, xin, yin, gamma):
     """Lorentzian broadening function
 
     Parameters
@@ -52,8 +54,6 @@ def lorentz_broaden(x, xin, yin, gamma, shift=0.0):
         The input spectrum.
     gamma : float
         Lorentzian broadening gamma.
-    shift : float
-        The energy shift.
 
     Returns
     -------
@@ -61,14 +61,12 @@ def lorentz_broaden(x, xin, yin, gamma, shift=0.0):
         Broadened spectrum.
     """
 
-    x1, x2 = np.meshgrid(x, xin + shift)
-
-    return (
-        np.dot(cauchy.pdf(x1, x2, gamma).T, yin) / len(xin) * (xin[-1] - xin[0])
-    )
+    x1, x2 = np.meshgrid(x, xin)
+    dx = xin[-1] - xin[0]
+    return np.dot(cauchy.pdf(x1, x2, gamma).T, yin) / len(xin) * dx
 
 
-def voigt_broaden(x, xin, yin, sigma, gamma, shift=0.0):
+def voigt_broaden(x, xin, yin, sigma, gamma):
     """Voigt broadening function
 
     Parameters
@@ -83,8 +81,6 @@ def voigt_broaden(x, xin, yin, sigma, gamma, shift=0.0):
         Voigt broadening sigma.
     gamma : float
         Voigt broadening gamma.
-    shift : float
-        The energy shift.
 
[…]
