### 2023-04-21 c000b03
add ocean anatase spectrum average over polarization directions
