+++ notebooks/spectra_files/anatase_ocean.txt
@@ -1,0 +1,2001 @@
+-0.20000000E+02 7.06363e-09
+-0.19950000E+02 7.07839e-09
+-0.19900000E+02 7.0932e-09
+-0.19850000E+02 7.10807e-09
+-0.19800000E+02 7.12298e-09
+-0.19750000E+02 7.13794e-09
+-0.19700000E+02 7.15295e-09
+-0.19650000E+02 7.16801e-09
+-0.19600000E+02 7.18312e-09
+-0.19550000E+02 7.19828e-09
+-0.19500000E+02 7.21349e-09
+-0.19450000E+02 7.22876e-09
+-0.19400000E+02 7.24407e-09
+-0.19350000E+02 7.25944e-09
+-0.19300000E+02 7.27486e-09
+-0.19250000E+02 7.29033e-09
+-0.19200000E+02 7.30585e-09
+-0.19150000E+02 7.32143e-09
+-0.19100000E+02 7.33705e-09
+-0.19050000E+02 7.35274e-09
+-0.19000000E+02 7.36847e-09
+-0.18950000E+02 7.38426e-09
+-0.18900000E+02 7.4001e-09
+-0.18850000E+02 7.416e-09
+-0.18800000E+02 7.43195e-09
+-0.18750000E+02 7.44796e-09
+-0.18700000E+02 7.46402e-09
+-0.18650000E+02 7.48013e-09
+-0.18600000E+02 7.49631e-09
+-0.18550000E+02 7.51253e-09
+-0.18500000E+02 7.52882e-09
+-0.18450000E+02 7.54516e-09
+-0.18400000E+02 7.56155e-09
+-0.18350000E+02 7.57801e-09
+-0.18300000E+02 7.59452e-09
+-0.18250000E+02 7.61109e-09
+-0.18200000E+02 7.62771e-09
+-0.18150000E+02 7.64439e-09
+-0.18100000E+02 7.66114e-09
+-0.18050000E+02 7.67794e-09
+-0.18000000E+02 7.6948e-09
+-0.17950000E+02 7.71172e-09
+-0.17900000E+02 7.72869e-09
+-0.17850000E+02 7.74573e-09
+-0.17800000E+02 7.76283e-09
+-0.17750000E+02 7.77999e-09
+-0.17700000E+02 7.79721e-09
+-0.17650000E+02 7.81449e-09
+-0.17600000E+02 7.83183e-09
+-0.17550000E+02 7.84923e-09
+-0.17500000E+02 7.8667e-09
+-0.17450000E+02 7.88422e-09
+-0.17400000E+02 7.90181e-09
+-0.17350000E+02 7.91947e-09
+-0.17300000E+02 7.93718e-09
+-0.17250000E+02 7.95496e-09
+-0.17200000E+02 7.9728e-09
+-0.17150000E+02 7.99071e-09
+-0.17100000E+02 8.00868e-09
+-0.17050000E+02 8.02672e-09
+-0.17000000E+02 8.04482e-09
+-0.16950000E+02 8.06299e-09
+-0.16900000E+02 8.08122e-09
+-0.16850000E+02 8.09952e-09
+-0.16800000E+02 8.11788e-09
+-0.16750000E+02 8.13631e-09
+-0.16700000E+02 8.15481e-09
+-0.16650000E+02 8.17338e-09
+-0.16600000E+02 8.19202e-09
+-0.16550000E+02 8.21072e-09
+-0.16500000E+02 8.22949e-09
+-0.16450000E+02 8.24833e-09
+-0.16400000E+02 8.26724e-09
+-0.16350000E+02 8.28622e-09
+-0.16300000E+02 8.30527e-09
+-0.16250000E+02 8.32439e-09
[…]
