### 2023-04-21 #5182 @larsenkg
@matthewcarbone I do not have any major concerns about this software or its suitability to JOSS. I am not in the XAS field, so some of the details of this software and its use cases are beyond me. The addition of the tutorial notebook was helpful. From what I do understand, this software will be a boon to your community. I only have a few thoughts relating to community guidelines and your references (and then some observations):

- I was not able to find "clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support". I recommend that you add a community guidelines section to your project README. The "Developers" subsection under the "Installation" section partially covers this. For example, the default assumption for bug reports is to use GitHub issues, but that could be made clear.
- The reference to Meng et al. (`xanesbench` in your bib file) is insufficient. I believe this is the same reference @maurov mentioned correcting.

My observations on installing for development and running tests:
- The install scripts did not work on Windows (using bash), possibly because of weirdness with /dev/stdin. I fixed the problem by using `xargs`, as below.
- The `test_from_materials_project` failed, even though PMG_API_KEY is set properly and worked in the notebook. (Also possibly a Windows error.)

If you would like me to file an issue for either of these, let me know.

```bash
install_doc_requirements_only () {
    python3 -c 'import toml; c = toml.load("pyproject.toml"); print(" ".join(c["project"]["optional-dependencies"]["doc"]))' | xargs pip install
}

install_test_requirements_only () {
    python3 -c 'import toml; c = toml.load("pyproject.toml"); print(" ".join(c["project"]["optional-dependencies"]["test"]))' | xargs pip install
}

install_requirements() {
    python3 -c 'import toml; c = toml.load("pyproject.toml"); print(" ".join(c["project"]["dependencies"]))' | xargs pip install
}
```
