+++ lightshow/_tests/test_broaden.py
@@ -2,7 +2,7 @@ import pytest
 
 import numpy as np
 
-from lightshow.postprocess import broaden
+from lightshow.postprocess.broaden import SiteSpectrum
 
 
 @pytest.fixture
@@ -14,106 +14,47 @@ def anatase_exp(spectra_file_directory):
 
 
 @pytest.fixture
-def anatase_theory(spectra_file_directory):
+def anatase_VASP_theory(spectra_file_directory):
     # Anatase TiO2 theoretical simulation is calculated by VASP
-    unbroaden_spect = np.loadtxt(spectra_file_directory / "anatase_theory.txt")
+    unbroaden_spect = np.loadtxt(
+        spectra_file_directory / "anatase_theory_VASP.txt"
+    )
 
     # choosing smaller range can make the optimization processs faster
-    xin = unbroaden_spect.T[0][20000:35000]
-    yin = unbroaden_spect.T[1][20000:35000]
+    xin = unbroaden_spect[30000:35000, :]
+    yin = unbroaden_spect[30000:35000, :]
     return [xin, yin]
 
 
-def test_broaden(anatase_theory, sample_file_directory):
-    # Output energy grid
-    x = np.linspace(4950, 5010, 1000, endpoint=True)
+def test_broaden(anatase_VASP_theory, anatase_exp, sample_file_directory):
+    """Makes sure the broadening code at least works!"""
 
-    # Create site spectrum object for broadening
-    ana = broaden.SiteSpectum(anatase_theory[0], anatase_theory[1], weight=1.0)
-
-    spect = broaden.gauss_broaden(
-        x, ana.spectrum[0], ana.spectrum[1], sigma=1.0, shift=100
-    )
-    assert (
-        max(
-            abs(
-                spect
-                - np.loadtxt(sample_file_directory / "gauss_broad.txt").T[1]
-            )
-        )
-        < 1e-10
-    )
-
-    spect = broaden.lorentz_broaden(
-        x, ana.spectrum[0], ana.spectrum[1], gamma=1.0, shift=100
-    )
-    assert (
-        max(
-            abs(
-                spect
-                - np.loadtxt(sample_file_directory / "lorentz_broad.txt").T[1]
-            )
-        )
-        < 1e-10
-    )
-
-    spect = broaden.voigt_broaden(
-        x, ana.spectrum[0], ana.spectrum[1], sigma=0.5, gamma=0.5, shift=100
-    )
-    assert (
-        max(
-            abs(
-                spect
-                - np.loadtxt(sample_file_directory / "voigt_broad.txt").T[1]
-            )
-        )
-        < 1e-10
-    )
-
-
-def test_paras_optimize(anatase_theory, anatase_exp, sample_file_directory):
-    # Collect XAS sumulation information
+    # Some information we'll need for the energy-dependent Voigt broadening
     TiK_core_state = -4864.0371
     TiK_core_hole_lifetime = 0.89
-
-    Ana_core_hole_fermi_energy = 5.1302
-
-    Fermi = Ana_core_hole_fermi_energy - TiK_core_state
+    ana_core_hole_fermi_energy = 5.1302
+    e_fermi = ana_core_hole_fermi_energy - TiK_core_state
 
     # Output energy grid
     x = np.linspace(4950, 5010, 1000, endpoint=True)
 
-    # Enter the information to site spectrum object
-    ana = broaden.SiteSpectum(
-        anatase_theory[0], anatase_theory[1], fermi=Fermi, weight=1.0
-    )
+    vasp_spectrum = SiteSpectrum(anatase_VASP_theory.copy(), e_fermi=e_fermi)
[…]
