### 2023-04-24 82e8e3e
add exciting and xpsectra references
