+++ notebooks/spectra_files/anatase_xspectra.txt
@@ -1,0 +1,400 @@
+-15.0    6.3e-07
+-14.78696742    6.3e-07
+-14.57393484    6.4e-07
+-14.36090226    6.6e-07
+-14.14786967    6.6e-07
+-13.93483709    6.6e-07
+-13.72180451    6.6e-07
+-13.50877193    6.9e-07
+-13.29573935    6.9e-07
+-13.08270677    6.9e-07
+-12.86967419    6.9e-07
+-12.6566416    7.199999999999999e-07
+-12.44360902    7.199999999999999e-07
+-12.23057644    7.199999999999999e-07
+-12.01754386    7.3e-07
+-11.80451128    7.5e-07
+-11.5914787    7.5e-07
+-11.37844612    7.5e-07
+-11.16541353    7.799999999999999e-07
+-10.95238095    7.799999999999999e-07
+-10.73934837    7.799999999999999e-07
+-10.52631579    8.100000000000001e-07
+-10.31328321    8.100000000000001e-07
+-10.10025063    8.100000000000001e-07
+-9.88721805    8.4e-07
+-9.67418546    8.4e-07
+-9.46115288    8.5e-07
+-9.2481203    8.699999999999999e-07
+-9.03508772    8.699999999999999e-07
+-8.82205514    9e-07
+-8.60902256    9e-07
+-8.39598997    9.3e-07
+-8.18295739    9.3e-07
+-7.96992481    9.5e-07
+-7.75689223    9.600000000000001e-07
+-7.54385965    9.8e-07
+-7.33082707    9.9e-07
+-7.11779449    1.01e-06
+-6.9047619    1.02e-06
+-6.69172932    1.05e-06
+-6.47869674    1.07e-06
+-6.26566416    1.08e-06
+-6.05263158    1.1e-06
+-5.839599    1.13e-06
+-5.62656642    1.14e-06
+-5.41353383    1.1600000000000001e-06
+-5.20050125    1.19e-06
+-4.98746867    1.22e-06
+-4.77443609    1.2499999999999999e-06
+-4.56140351    1.2700000000000001e-06
+-4.34837093    1.3e-06
+-4.13533835    1.33e-06
+-3.92230576    1.38e-06
+-3.70927318    1.4099999999999998e-06
+-3.4962406    1.46e-06
+-3.28320802    1.51e-06
+-3.07017544    1.5599999999999999e-06
+-2.85714286    1.6200000000000002e-06
+-2.64411028    1.69e-06
+-2.43107769    1.77e-06
+-2.21804511    1.8400000000000002e-06
+-2.00501253    1.96e-06
+-1.79197995    2.08e-06
+-1.57894737    2.2399999999999997e-06
+-1.36591479    2.47e-06
+-1.15288221    2.77e-06
+-0.93984962    3.2299999999999996e-06
+-0.72681704    4e-06
+-0.51378446    5.6100000000000005e-06
+-0.30075188    1.007e-05
+-0.0877193    6.031e-05
+0.12531328    0.00012486000000000002
+0.33834586    5.443e-05
+0.55137845    1.3320000000000001e-05
+0.76441103    1.102e-05
+0.97744361    3.64e-05
+1.19047619    0.00041757
+1.40350877    0.00043209
+1.61654135    0.00065198
+1.82957393    0.00142238
+2.04260652    0.00054614
+2.2556391    0.00027342
+2.46867168    0.0007064700000000001
+2.68170426    0.00065859
+2.89473684    0.00023052
+3.10776942    0.00021394
+3.32080201    0.00054422
+3.53383459    0.00031311999999999996
+3.74686717    0.00028105999999999997
+3.95989975    0.00102768
+4.17293233    0.00052718
[…]
