### 2023-04-24 b3c99c8
add exciting and xpsectra references
