+++ notebooks/spectra_files/anatase_exciting.txt
@@ -1,0 +1,1000 @@
+4849.068603852  7.412236753674784e-08
+4849.131190034877  7.443082752105754e-08
+4849.193776217756  7.474154527281904e-08
+4849.256362400634  7.505454703409193e-08
+4849.318948583512  7.536985949823563e-08
+4849.38153476639  7.568750982059971e-08
+4849.444120949268  7.600752562953998e-08
+4849.506707132145  7.632993503781045e-08
+4849.569293315024  7.665476665427245e-08
+4849.631879497902  7.698204959596897e-08
+4849.69446568078  7.731181350061515e-08
+4849.757051863658  7.764408853944315e-08
+4849.819638046536  7.797890543048094e-08
+4849.882224229414  7.831629545223168e-08
+4849.944810412292  7.865629045783728e-08
+4850.007396595171  7.899892288966418e-08
+4850.069982778048  7.934422579436713e-08
+4850.132568960926  7.969223283848763e-08
+4850.195155143804  8.004297832453232e-08
+4850.257741326682  8.03964972075883e-08
+4850.320327509559  8.07528251125378e-08
+4850.382913692439  8.111199835181751e-08
+4850.445499875316  8.147405394378331e-08
+4850.508086058194  8.18390296317493e-08
+4850.570672241072  8.220696390364412e-08
+4850.63325842395  8.257789601235458e-08
+4850.695844606827  8.295186599682634e-08
+4850.758430789706  8.332891470386516e-08
+4850.821016972584  8.370908381074511e-08
+4850.883603155462  8.409241584860072e-08
+4850.94618933834  8.447895422671486e-08
+4851.008775521218  8.486874325764966e-08
+4851.071361704096  8.526182818330216e-08
+4851.133947886974  8.565825520196889e-08
+4851.196534069853  8.605807149637758e-08
+4851.25912025273  8.646132526277353e-08
+4851.321706435608  8.6868065741154e-08
+4851.384292618486  8.727834324661291e-08
+4851.446878801364  8.769220920189315e-08
+4851.509464984242  8.810971617124837e-08
+4851.572051167121  8.853091789558377e-08
+4851.634637349998  8.895586932898295e-08
+4851.697223532876  8.938462667673414e-08
+4851.759809715754  8.981724743483283e-08
+4851.822395898632  9.025379043108047e-08
+4851.88498208151  9.069431586790631e-08
+4851.947568264388  9.11388853668926e-08
+4852.010154447266  9.15875620151772e-08
+4852.072740630144  9.204041041375921e-08
+4852.135326813021  9.249749672789478e-08
+4852.1979129959  9.29588887395816e-08
+4852.260499178778  9.342465590228853e-08
+4852.323085361656  9.389486939810137e-08
+4852.385671544534  9.43696021973039e-08
+4852.448257727412  9.484892912057375e-08
+4852.51084391029  9.533292690398791e-08
+4852.573430093168  9.582167426687719e-08
+4852.636016276046  9.631525198273814e-08
+4852.698602458924  9.681374295342422e-08
+4852.761188641803  9.731723228668197e-08
+4852.82377482468  9.782580737727599e-08
+4852.886361007558  9.833955799195479e-08
+4852.948947190435  9.885857635835658e-08
+4853.011533373314  9.938295725817637e-08
+4853.074119556192  9.9912798124762e-08
+4853.13670573907  1.0044819914549765e-07
+4853.199291921948  1.0098926336912701e-07
+4853.261878104826  1.0153609681836505e-07
+4853.324464287703  1.020888086081718e-07
+4853.387050470582  1.0264751106990137e-07
+4853.44963665346  1.0321231988174161e-07
[…]
