### 2023-04-24 db6f9a6
Replace 1/ld with alpha and update the broadening noteboook
