### 2023-04-25 edc5375
fix the test code
