+++ docs/source/index.rst
@@ -7,15 +7,43 @@
    modules
    project
 
-.. include:: ../../README.rst
-   :start-after: inclusion-marker-LIGHTSHOW-begin
-   :end-before: inclusion-marker-LIGHTSHOW-end
+.. image:: https://raw.githubusercontent.com/AI-multimodal/Lightshow/master/docs/_static/images/lightshow.jpg
+    :align: right
+    :alt: sysfs line plot
+    :width: 1000px
+
+.. image:: https://github.com/AI-multimodal/Lightshow/actions/workflows/ci.yml/badge.svg
+ :target: https://github.com/AI-multimodal/Lightshow/actions/workflows/ci.yml
+
+.. image:: https://codecov.io/gh/AI-multimodal/Lightshow/branch/master/graph/badge.svg?token=CW7BMFA5O7 
+ :target: https://codecov.io/gh/AI-multimodal/Lightshow
+ 
+.. image:: https://app.codacy.com/project/badge/Grade/d31a4e18672c4d71bbaafa719181c140
+ :target: https://www.codacy.com/gh/AI-multimodal/Lightshow/dashboard?utm_source=github.com&amp;utm_medium=referral&amp;utm_content=AI-multimodal/Lightshow&amp;utm_campaign=Badge_Grade
+ 
+.. image:: https://joss.theoj.org/papers/a9cabcd7f4b85a926a797997c6622b43/status.svg
+  :target: https://joss.theoj.org/papers/a9cabcd7f4b85a926a797997c6622b43
+
+.. image:: https://zenodo.org/badge/DOI/10.48550/arXiv.2211.04452.svg
+   :target: https://doi.org/10.48550/arXiv.2211.04452
+
+=============
+
+**Lightshow** is a Python library for easily generating computational spectroscopy input files.
 
 Overview
 ========
 
-.. include:: ../../README.rst
-   :start-after: inclusion-marker-LIGHTSHOW-overview-begin
-   :end-before: inclusion-marker-LIGHTSHOW-overview-end
+Often, it can be daunting to create comprehensive, well documented databases of materials structures and their x-ray absorption spectra. **Lightshow** solves this problem, allowing new users to choose sensible defaults for their calculations, while simultaneously exposing all functionality for experts.
+
+**Lightshow** aims to provide a "one-stop-shop" for input file generation, and currently supports the following codes:
+
+- FEFF
+- VASP
+- OCEAN
+- EXCITING
+- Xspectra
+
+with more on the way! The software is intended to be user-friendly, extensively documented and tested, and extendable for those users who wish to add additional spectroscopy functionalities. There are also a few comprehensive tutorials to help you get started.
 
 
