+++ docs/source/project/funding.rst
@@ -1,6 +1,33 @@
 Funding
 =======
 
-.. include:: ../../../README.rst
-   :start-after: inclusion-marker-LIGHTSHOW-funding-begin
-   :end-before: inclusion-marker-LIGHTSHOW-funding-end
+This research is based upon work supported by the U.S. Department of
+Energy, Office of Science, Office Basic Energy Sciences, under Award
+Number FWP PS-030. This research used resources of the Center for
+Functional Nanomaterials (CFN), which is a U.S. Department of Energy
+Office of Science User Facility, at Brookhaven National Laboratory under
+Contract No. DE-SC0012704.
+
+The Software resulted from work developed under a U.S. Government
+Contract No. DE-SC0012704 and are subject to the following terms: the
+U.S. Government is granted for itself and others acting on its behalf a
+paid-up, nonexclusive, irrevocable worldwide license in this computer
+software and data to reproduce, prepare derivative works, and perform
+publicly and display publicly.
+
+THE SOFTWARE IS SUPPLIED \"AS IS\" WITHOUT WARRANTY OF ANY KIND. THE
+UNITED STATES, THE UNITED STATES DEPARTMENT OF ENERGY, AND THEIR
+EMPLOYEES: (1) DISCLAIM ANY WARRANTIES, EXPRESS OR IMPLIED, INCLUDING
+BUT NOT LIMITED TO ANY IMPLIED WARRANTIES OF MERCHANTABILITY, FITNESS
+FOR A PARTICULAR PURPOSE, TITLE OR NON-INFRINGEMENT, (2) DO NOT ASSUME
+ANY LEGAL LIABILITY OR RESPONSIBILITY FOR THE ACCURACY, COMPLETENESS, OR
+USEFULNESS OF THE SOFTWARE, (3) DO NOT REPRESENT THAT USE OF THE
+SOFTWARE WOULD NOT INFRINGE PRIVATELY OWNED RIGHTS, (4) DO NOT WARRANT
+THAT THE SOFTWARE WILL FUNCTION UNINTERRUPTED, THAT IT IS ERROR-FREE OR
+THAT ANY ERRORS WILL BE CORRECTED.
+
+IN NO EVENT SHALL THE UNITED STATES, THE UNITED STATES DEPARTMENT OF
+ENERGY, OR THEIR EMPLOYEES BE LIABLE FOR ANY DIRECT, INDIRECT,
+INCIDENTAL, CONSEQUENTIAL, SPECIAL OR PUNITIVE DAMAGES OF ANY KIND OR
+NATURE RESULTING FROM EXERCISE OF THIS LICENSE AGREEMENT OR THE USE OF
+THE SOFTWARE.
