### 2023-06-23 #5182 @matthewcarbone
Ok so my required todo checklist as it stands:

- [x] Add more example notebooks
- [ ] [Add sentences to the manuscript referencing that the code is part of a larger workflow](https://github.com/openjournals/joss-reviews/issues/5182#issuecomment-1498759674)
- [ ] Correct citation syntax for [Meng et al.](https://doi.org/10.48550/arXiv.2303.17089)
- [ ] Add citation for [Pymatgen](https://doi.org/10.1016/j.commatsci.2012.10.028)
- [ ] Add a citation for [FEFF10/Corvus](https://journals.iucr.org/s/issues/2021/06/00/hf5422/)
- [ ] [Add to the documentation how one could contribute for adding support for additional calculators](https://github.com/openjournals/joss-reviews/issues/5182#issuecomment-1442158295)
- [ ] [Update the installation instructions with more detail](https://github.com/openjournals/joss-reviews/issues/5182#issuecomment-1505563873)
- [ ] [Update instructions on how to manually run tests after local installation](https://github.com/openjournals/joss-reviews/issues/5182#issuecomment-1505563873)
- [ ] [Add community guidelines section to README](https://github.com/openjournals/joss-reviews/issues/5182#issuecomment-1518159044)

Other possible todos:

- [ ] [Examine possible Windows-specific problems with installation](https://github.com/openjournals/joss-reviews/issues/5182#issuecomment-1518159044)
