+++ paper.bib
@@ -97,9 +97,10 @@
 }
 
 @article{xanesbench,
-  author={Fanchen Meng and Benedikt Maurer and Fabian Peschel and Sencer Selcuk and  Mark Hybertsen and Xiaohui Qu and Christian Vorwerk and Claudia Draxl and John Vinson and Deyu Lu},
-  note={private communications},
-  year={2022}
+  title={Multi-code Benchmark on Simulated Ti K-edge X-ray Absorption Spectra of Ti-O Compounds},
+  author={Meng, Fanchen and Maurer, Benedikt and Peschel, Fabian and Selcuk, Sencer and Hybertsen, Mark and Qu, Xiaohui and Vorwerk, Christian and Draxl, Claudia and Vinson, John and Lu, Deyu},
+  journal={arXiv preprint arXiv:2303.17089},
+  year={2023}
 }
 
 @book{de2008core,
