+++ paper.bib
@@ -8,6 +8,17 @@
     year = {2013}
 }
 
+@article{jain2011high,
+  title={A high-throughput infrastructure for density functional theory calculations},
+  author={Jain, Anubhav and Hautier, Geoffroy and Moore, Charles J and Ong, Shyue Ping and Fischer, Christopher C and Mueller, Tim and Persson, Kristin A and Ceder, Gerbrand},
+  journal={Computational Materials Science},
+  volume={50},
+  number={8},
+  pages={2295--2310},
+  year={2011},
+  publisher={Elsevier}
+}
+
 @article{rehr2010parameter,
   title={Parameter-free calculations of X-ray spectra with FEFF9},
   author={Rehr, John J and Kas, Joshua J and Vila, Fernando D and Prange, Micah P and Jorissen, Kevin},
