+++ paper.bib
@@ -120,3 +120,14 @@
   year={2008},
   publisher={CRC press}
 }
+
+@article{kas2021advanced,
+  title={Advanced calculations of X-ray spectroscopies with FEFF10 and Corvus},
+  author={Kas, JJ and Vila, FD and Pemmaraju, CD and Tan, TS and Rehr, JJ},
+  journal={Journal of Synchrotron Radiation},
+  volume={28},
+  number={6},
+  pages={1801--1810},
+  year={2021},
+  publisher={International Union of Crystallography}
+}
