### 2023-06-29 f61bec8
Create CONTRIBUTING.md
