+++ CONTRIBUTING.md
@@ -1,0 +1,12 @@
+# Contributing
+
+We welcome any and all contributions to Lightshow! In particular, we highly encourage incorporating support for a new code.
+If you wish to contribute, please feel free to do the following:
+
+1. Create a new GitHub [Issue](https://github.com/AI-multimodal/Lightshow/issues) detailing what you would like to do (optional).
+2. Fork your own copy of Lightshow, [make your changes](https://github.com/firstcontributions/first-contributions) and open a [Pull Request](https://github.com/AI-multimodal/Lightshow/pulls).
+3. Discuss with the developers and other community members, refining your code as such.
+4. Ensure all tests pass!
+5. Once everyone is satisfied, we will merge your code and make a new release.
+
+## Bug reporting
