### 2023-06-29 1215bad
Add bug reporting section to contributing
