+++ CONTRIBUTING.md
@@ -10,3 +10,9 @@ If you wish to contribute, please feel free to do the following:
 5. Once everyone is satisfied, we will merge your code and make a new release.
 
 ## Bug reporting
+
+If you find an issue or bug with Lightshow, please report it by opening a
+a new GitHub [Issue](https://github.com/AI-multimodal/Lightshow/issues)
+detailing the problem. Please make sure your problem is reproducible, and that
+you document important details such as the package versions you're using (e.g.
+the version of Lightshow, Pymatgen, etc.) and your operating system.
