### 2023-07-01 1d40a54
Add instructions on how to add new parameters objects
