### 2023-07-01 bd7aabb
Add instructions for running tests locally
