### 2023-07-01 4887bdc
Use more specific instructions in readme for tests
