### 2023-07-01 #5182 @matthewcarbone
@ppxasjsm ok all done!

@larsenkg since it was your idea, do you want to open a PR to fix the Windows-related install? 👍
