### 2023-07-04 #5182 @ppxasjsm
I have gone through the paper. Here is a small list of corrections from me:

- [ ] Can you make sure that X-ray is written as X-ray everywhere?
- [ ] Frist-principles should be hyphenated consistently.
- [ ] The emergence of the high-performance computing hardware architecture-> The emergence of high-performance computing hardware architecture
- [ ] continue to fuel the advance of first-principles simulations in materials design. -> continues to fuel the advances of first-principles simulations in materials design.
- [ ] ultraviolet–visible uses the wrong dash I believe. Please double check.
- [ ] Finally, we also note that Lightshow is ... use `Lightshow` here too!

Can you please address these, double check you are happy with everything and then we can proceed to recommend accept. For this step you will need to do the following things:
