### 2023-07-05 #5182 @matthewcarbone
Keeping my own copy so I can keep track of progress:

> I have gone through the paper. Here is a small list of corrections from me:
>
> * [ ]  Can you make sure that X-ray is written as X-ray everywhere?
> * [ ]  Frist-principles should be hyphenated consistently.
> * [ ]  The emergence of the high-performance computing hardware architecture-> The emergence of high-performance computing hardware architecture
> * [ ]  continue to fuel the advance of first-principles simulations in materials design. -> continues to fuel the advances of first-principles simulations in materials design.
> * [ ]  ultraviolet–visible uses the wrong dash I believe. Please double check.
> * [ ]  Finally, we also note that Lightshow is ... use `Lightshow` here too!
>
> Can you please address these, double check you are happy with everything and then we can proceed to recommend accept. For this step you will need to do the following things:
>
> * [ ]  Make a tagged release of your software, and list the version tag of the archived version here.
> * [ ]  Archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional repository)
> * [ ]  Check the archival deposit (e.g., in Zenodo) has the correct metadata. This includes the title (should match the paper title) and author list (make sure the list is correct and people who only made a small fix are not on it). You may also add the authors' ORCID.
> * [ ]  Please list the DOI of the archived version here.
