### 2023-07-05 cb87210
Implement editorial suggestions
