### 2023-07-06 #5182 @matthewcarbone
@ppxasjsm all done. Note for UV-vis apparently JOSS doesn't render the triple dash correctly. I think it's fine as is!
