### 2023-07-12 #5182 @ppxasjsm
Hi @matthewcarbone, I have one last thing for you. Can you please add the DOI: https://doi.org/10.1016/j.commatsci.2011.02.023 to "A high-throughput infrastructure for density functional theory calculations" by Anhubhav et al. Thanks!
