+++ paper.bib
@@ -16,7 +16,8 @@
   number={8},
   pages={2295--2310},
   year={2011},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.commatsci.2011.02.023}
 }
 
 @article{rehr2010parameter,
