### 2023-07-25 #5182 @Kevin-Mattheus-Moerman
@matthewcarbone I am the AEiC for this track and here to help with the final steps. I have reviewed the repository, the archive link, the paper, and this review thread. Most seems in order. Below are some points that do require your attention.

- [ ] Please add contributing guidelines. It is recommended to have a dedicated `CONTRIBUTING.md` file, and to refer to it in the `README.md` as well. The file content can be quite basic/simple if you like. There are many sites offering examples/templates for this (just saw this site too which might help: https://contributing.md/example/). Let me know if you have questions.

Some references need formatting/updating.
Our system did not detect these (@openjournals/dev @xuanxu why not actually?) but there seem to be many DOIs that are missing, below are some examples, but please check all your references and add DOIs where possible.
- [ ] _Effects of electron-phonon coupling on absorption spectrum: K edge of hexagonal boron nitride_: https://doi.org/10.1103/PhysRevB.98.235205
- [ ] _Advanced calculations of X-ray spectroscopies with FEFF10 and Corvus_ : https://doi.org/10.1107/S1600577521008614
- [ ] _Projector augmented wave calculation of x-ray absorption spectra at the L2,3 edges_ : https://doi.org/10.1103/PhysRevB.87.205105
- [ ] _First-principles calculations of x-ray absorption in a scheme based on ultrasoft pseudopotentials_ : https://doi.org/10.1103/PhysRevB.80.075102
- [ ]  _Multi-code Benchmark on Simulated Ti K-edge X-ray Absorption Spectra of Ti-O Compounds_ : https://doi.org/10.48550/arXiv.2303.17089
- [ ] _Advances in the OCEAN-3 spectroscopy package_ : https://doi.org/10.1039/D2CP01030E
- [ ] This reference seems to be incomplete in terms of the title and the DOI:
```
Jain, A., Ong, S. P., Hautier, G., Chen, W., Richards, W. D., Dacek, S., Cholia, S., Gunter,
D., Skinner, D., Ceder, G., & Persson, K. a. (2013). APL Materials, 1(1), 011002.
```
https://doi.org/10.1063/1.4812323

When you've completed the above you can update the draft and check the references again using `@editorialbot check references`.
