+++ paper.bib
@@ -1,11 +1,13 @@
 @article{Jain2013,
-    author = {Jain, Anubhav and Ong, Shyue Ping and Hautier, Geoffroy and Chen, Wei and Richards, William Davidson and Dacek, Stephen and Cholia, Shreyas and Gunter, Dan and Skinner, David and Ceder, Gerbrand and Persson, Kristin a.},
+    author = {Jain, Anubhav and Ong, Shyue Ping and Hautier, Geoffroy and Chen, Wei and Richards, William Davidson and Dacek, Stephen and Cholia, Shreyas and Gunter, Dan and Skinner, David and Ceder, Gerbrand and Persson, Kristin A.},
+    title = {Commentary: The Materials Project: A materials genome approach to accelerating materials innovation},
     issn = {2166532X},
     journal = {APL Materials},
     number = {1},
     pages = {011002},
     volume = {1},
-    year = {2013}
+    year = {2013},
+    doi = {10.1063/1.4812323}
 }
 
 @article{jain2011high,
@@ -51,7 +53,8 @@
   number={7},
   pages={075102},
   year={2009},
-  publisher={APS}
+  publisher={APS},
+  doi={10.1103/PhysRevB.80.075102}
 }
 
 @article{bunuau2013projector,
@@ -62,7 +65,8 @@
   number={20},
   pages={205105},
   year={2013},
-  publisher={APS}
+  publisher={APS},
+  doi={10.1103/PhysRevB.87.205105}
 }
 
 @article{vinson2011bethe,
@@ -84,7 +88,8 @@
   volume={24},
   pages={12787},
   year={2022},
-  publisher={RSC}
+  publisher={RSC},
+  doi={10.1039/D2CP01030E}
 }
 
 @article{vasp-xas,
@@ -95,7 +100,8 @@
   number={23},
   pages={235205},
   year={2018},
-  publisher={APS}
+  publisher={APS},
+  doi={10.1103/PhysRevB.98.235205}
 }
 
 @article{exciting,
@@ -112,7 +118,8 @@
   title={Multi-code Benchmark on Simulated Ti K-edge X-ray Absorption Spectra of Ti-O Compounds},
   author={Meng, Fanchen and Maurer, Benedikt and Peschel, Fabian and Selcuk, Sencer and Hybertsen, Mark and Qu, Xiaohui and Vorwerk, Christian and Draxl, Claudia and Vinson, John and Lu, Deyu},
   journal={arXiv preprint arXiv:2303.17089},
-  year={2023}
+  year={2023},
+  doi={10.48550/arXiv.2303.17089}
 }
 
 @book{de2008core,
@@ -130,5 +137,6 @@
   number={6},
   pages={1801--1810},
   year={2021},
-  publisher={International Union of Crystallography}
+  publisher={International Union of Crystallography},
+  doi={10.1107/S1600577521008614}
 }
