### 2023-07-25 #5182 @matthewcarbone
@Kevin-Mattheus-Moerman we do already have a `CONTRIBUTING.md` file in the root directory of the project. Note that the `joss-manuscript` branch is completely standalone and that the main project is on `master`. Are the current contributing guidelines sufficient?
