### 2023-07-28 #5182 @Kevin-Mattheus-Moerman
@matthewcarbone currently your `paper.bib` file says:
![Screenshot from 2023-07-28 09-45-30](https://github.com/openjournals/joss-reviews/assets/8392709/7d31c07e-f91e-45dd-ad69-136a5858fdd7)

Hence the above updated paper draft still lacks the DOI:

![Screenshot from 2023-07-28 09-45-01](https://github.com/openjournals/joss-reviews/assets/8392709/5a563f44-7416-4787-aad9-4a47043af1fe)

Perhaps you were working in a different branch? Can you update your [bib file](https://github.com/AI-multimodal/Lightshow/blob/joss-manuscript/paper.bib) and the paper in the `joss-manuscript` branch?
