+++ paper.bib
@@ -42,7 +42,8 @@
   number={19},
   pages={195107},
   year={2002},
-  publisher={APS}
+  publisher={APS},
+  doi={10.1103/PhysRevB.66.195107}
 }
 
 @article{gougoussis2009first,
